Add API call @ componentDidMount that gets reviews
Added an initial API call at componentDidMount lifecycle of parent App component that gets all the reviews for the currentGameId in the parent App component's state. This involved setting up a DB dir and creating a function that queries the DB for all the reviews pertaining to a game ID upon receving the get request with the gameId query paramenter.

LAT-49